+++ commonmark/src/main/java/org/commonmark/internal/inline/AutolinkInlineParser.java
@@ -19,7 +19,7 @@ public class AutolinkInlineParser implements InlineContentParser {
 
     private static final Pattern EMAIL =
             Pattern.compile(
-                    "^([a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*)$");
+                    "^[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*+$");
 
     @Override
     public ParsedInline tryParse(InlineParserState inlineParserState) {